+++ android/2023-emmsale/app/src/main/java/com/emmsale/presentation/ui/editMyProfile/recyclerView/ActivityDiffUtil.kt
@@ -1,0 +1,16 @@
+package com.emmsale.presentation.ui.editMyProfile.recyclerView
+
+import androidx.recyclerview.widget.DiffUtil
+import com.emmsale.data.model.Activity
+
+class ActivityDiffUtil : DiffUtil.ItemCallback<Activity>() {
+    override fun areItemsTheSame(
+        oldItem: Activity,
+        newItem: Activity,
+    ): Boolean = oldItem.id == newItem.id
+
+    override fun areContentsTheSame(
+        oldItem: Activity,
+        newItem: Activity,
+    ): Boolean = oldItem == newItem
+}
